requirements.txt
apache-airflow-providers-MySQL
pymysql
apache-airflow
pandas
sqlalchemy
pymysql
streamlit